requirements.txt
ansible==2.6.2
pycountry
authorizenet==1.1.3
bok_choy==0.8.0
bs4==0.0.1
celery==4.2.0
celery-haystack==0.10
coverage==4.5.3
python-dateutil==2.8.0
ddt==1.1.1
Django==2.2.1
django-avatar==4.1.0
django-otp==0.4.3
django-celery-email==2.0.0
django-haystack==2.8.1
django-hitcount==1.3.1
django-ipware==2.0.1
djangorestframework==3.9.1
djangorestframework-bulk==0.2.1
django-rest-recaptcha==1.0.1
django-reversion==3.0.4
django_money==0.14
django-premailer==0.2.0
django-custom-user==0.7
django-sendgrid-v5==0.8.0
django-signal-disabler==0.1.1
dwollav2==1.2.4
easy_thumbnails==2.5
factory_boy==2.8.1
freezegun==0.3.12
ipython==6.2.1
mailchimp3==3.0.5
markdown==2.6.11
pandas==0.23.3
PyVirtualDisplay==0.2.1
psycopg2>=2.7,<2.8 --no-binary psycopg2
python-telegram-bot==10.1.0
pytz>=2018.3
requests==2.20.0
selenium-requests==1.3
tblib==1.3.2
luhn==0.2.0
django-webpack-loader>=0.5.0
whoosh==2.7.4
qrcode==6.1
django-markdown2==0.3.1
git+https://github.com/Artconomy/short_stuff.git@0.1.2#egg=short_stuff
